Added new confirmation model box

+++ apps/host-ui/src/app/messages/messages.component.ts
@@ -1,3 +1,4 @@
+import { ConfirmService } from "./../_services/confirm.service";
 import { MessageService } from "./../_services/message.service";
 import { Component, OnInit } from '@angular/core';
 import { Message } from '../_models/message';
@@ -16,7 +17,7 @@ export class MessagesComponent implements OnInit {
 	container = 'Unread';
 	loading = false;	
 
-	constructor(private messgeService: MessageService) {}
+	constructor(private messgeService: MessageService, private confirmService: ConfirmService) {}
 
 	ngOnInit(): void {
 		this.loadMessages();
@@ -41,11 +42,15 @@ export class MessagesComponent implements OnInit {
 	  }
 
 	  deleteMessage(id: number) {
-		this.messgeService.deleteMessage(id).subscribe({
-			next: () => {
-				/// remove the deleted message from the ui
-				this.messages.splice(this.messages.findIndex(m => m.id === id), 1);
+		this.confirmService.confirm('Confirm delete message', 'This cannot be undone').subscribe(result => {
+			if(result) {
+				this.messgeService.deleteMessage(id).subscribe({
+					next: () => {
+						/// remove the deleted message from the ui
+						this.messages.splice(this.messages.findIndex(m => m.id === id), 1);
+					}
+				})
 			}
-		})
+		})		
 	  }
 }
